Possible win32 fix at high load.
Win32 may return a WSAENOBUFS (yes the ENOBUFS is missing) on send which would indicate
memory starvation within the OS. It is unclear whether that is a significant error but
for now we should treat it as temporary and retry the client later.   The docs are not
clear on whether ret is also set is an error is returned.