source client hijack fixes.
prevent source_init from being run when hijack is triggered, this prevents a bunch of
stats being initialised.

mpeg alignment was disabled as the mpeg data structure was not being transferred to new
client, this in turn caused FLV wrapping to get messed up.

The read bytes is also copied and reset when the old client is about to go, that way the
access log has reasonable values.