fix for very rare auth thread shutdown leaving client inactive case
auth threads are started as required up to a certain allowed number. While the
startup of an auth thread occurs in a lock, the shutdown did not.  As there is
typically several possible threads possible, the chance that all are at the same
point is remote, so is really a rare case.

We still take/release locks in the loop so we do not prevent new clients being
added, we just keep the lock until the thread handle is cleared.