build fix for cygwin
winsock2.h is not wanted on cygwin even though is does exist so make sure the
mingw32 requirement for it is only done for mingw32.  It also turns out that the
pthread lib on cygwin is missing the timed versions of the lock routines so I
have wrapped those in conditional compiled code.