stream stats update by slave thread reduced
in 2.3.2, the relays did not always have clients or other processing contexts to
handle their own stats so relied on the slave thread to update them which had a
side effect of updating more than what was required.  In the kh build, all the
sources (including relays) have a client on a worker so can do processing. This
now means the slave thread only needs to take care of virtual mountpoints ie
unavailable streams with an active fallback.

This should reduce the number of stats processed and prevent any racing between
the slave and worker threads.