Drop the FLV buffer expansion code.
Originally this was needed as we copied stream data into it for sending to
listeners but now we only store the FLV headers. The also fixes a possible
corruption issue if the block is relocated invalidating the vectors that had
previously been populated. Any subsequent restarting of a short write (which
could maniputate the vector list) could of sent junk data or could corrupt
other memory allocations like queue buffers.