fix for fallback over multiple mount.
The move listener routine was jumping out if no source was present, but should of
taken into account that mount details could refer to other mountpoints. This affected
cases eg A->B->C where B was no reserved.

The actual fix is in auth.c but to handle the common case I perform a lookup in set
fallback so that the cascade is not done for each listener.