hijack source fixes.
actually remove the tie-in of the old client and source in the swap routine as
we could race if not. This means we do not need the HIJACK flag anymore. Also make
sure the source lock is taken before checking, that way we prevent a race case in
there. Thw format specific swap routine also drops the surplus data if any as that
will contain incomplete frames that will never complete.