allow more tolerance on resync of streams
in certain cases, a resync of an existing stream could find a series of
matching markers leading to a quick trigger of the termination of the source.

This makes the routine for finding the sync marker work hard if there are some
sync bytes to look for.  This prevents matching for 0xFF only which could be
common enough to increase the resync counter to termination.  Now we match all
sync bytes, looping for more before returning.