mpeg parse fixup of the day. Corner case fix but needed more work
The previous code work fairly well but didn't handle the situation where
certain bits in the frame headers could of changed and invalidated the stream.

The main case demonstrating this was a non-encoding source client where the
files varied in whether CRC was enabled or not.  So instead of just matching
the first (upto) 3 bytes, I store a bitmask and match the intial frame value
against subsequent frame header after mask is applied. The mask allow for
certain bit to always match and others never.

Many just have re-encoded content which does not expose the issue. I've made
the resync counter based on bytes skipped, set to 20k and I've avoided some
odd case where finding a frame could come across many markers which are false
but trigger a number of memory copies.