win32 fix for timezone bug in timestamps
When msvc was used, Ed made a routine to make a datestamp with timezone offset
as the win32 api defaults to using a name. I had disabled this for logs when the
switch to mingw32 happened but the name is still an issue as it may be in a non-utf8
character set.   While this is not a problem for ithe access log, this becomes an
issue for the stats as that requires utf8 strings.

I've now moved the routine Ed did to util, cleaned it up to reduce some work and
used it via stats and logging.