force termination if no sockets can be allocated initially
For non-root starts, a failure to get any listening sockets initially
meant the server ran. This may be confusing for some especially if they
had already started icecast and it's on the expected port.  This may
also confuse centovacast.

If no listening sockets are available over reload then the server still
continues as existing listeners can be serviced and still gives chance
to recover if the situation was just down to a typo in the xml.