Add <x-forwarded-for> into <paths> for IP substitution
allows you to nominate certain addresses where reverse proxies can be
used. These trusted proxies can set the X-Forwarded-For header to specify
the real IP of the listener.

We don't allow them generally as header injection is too easy to trick
but usually this is for trusted load balancing proxies.