reduce reschedule time of readers if only small blocks arriving
some setups seem to only receive small writes eg less than 300 bytes,
which causes another reschedule for checking again. With this I'll try
to increase the wait time if only a small amount arrives.

This is experimental but should be fairly low impact in the normal case
but may need tweaking.