worker balance tolerance changes
only check whether listener needs to change worker every few seconds instead
of each time, as each time requires secondary locking, and change the point
at which the move is not considered.  This should force listeners to be on
the same worker as the source but may delay it if there is a chance the
source will move again.