Fixup listener switching to source client worker.
listener moves to the other worker provided moving there won't be too many
clients on that worker.  Having listeners on the same worker as the source
eases contention issues but a single stream with many listeners would max
out a worker fairly easy, so we only do this if there are no more than 1000
clients difference.  source clients can also move so this should allow for
a reasonable balance (hope!).

The balance is more like we had before, but now the rwlocks should have less
impact when listeners are spread out over workers.