worker balance work.
The main bits here are to allow for clients to switch workers if the load
differences is worth it.  This means that listeners will move to the source
client worker if not on already but also that it will move away if there is
sufficiently less.  The levels to prevent them happening are initial guesses

To help the routines, a number of underlying changes are done. The least busy
worker is maintained in a global and only updated from the slave thread each
second, saves multiple scanning.   Each worker will attempt to move N clients
when notified by the slave thread, which cycles through the workers 1 per
second.