internal unlock move.
up to now, the main source unlock (now rwlock) was done inside source_read although it
was taken in the parent finction. We now move the unlock to that parent function which
allows us to drop to the shutdown and cleanup parts without rescheduling the client.

The rescheduling meant that there was a window of opportunity for listeners to get in
and see that the source was inactive and put themselves into a waiting state delaying
shutdown until they awoke again. Now the shutdown is activated before the lock is
dropped.