crash fix, avoid invalidating source stats
issue #22 had a file with same name as stream getting used and the last listener
was triggering a free up of the stats, invalidating one used by source of the same
name that had connected afterwards.

because fallback to file can also cause a name clash with a stream reconnecting,
I've decided to put those stats under a name of "file-/mount", that way there is
no clash and that also avoids being invalidated.