switch to using pread for file reading, fallback/intro.
In cases where pread is missing, use a wrapper function that uses a lock.
win32 could use the overlapped io option for that but not in this commit.

The use of pread means there is no race issue with file access and rwlocks
and non streamed files don't take mutexes now. fallback files still do as
stats are generated