reduce mutex locking on throttled files, should help scaling across workers.
The lock was originally taken on each pass as a file pointer is not thread safe, nor
was the stats or bitrate updates. A File descriptor is now used with pread and bitrate
updates us there own spinlock so this deals with the stats updating which only done
every few seconds. This may become a spin lock later.