only reduce skip duration when data is available for queueing
It is possible for source clients with small writes to trigger a frequent
reschedule by the client which only reads the small amount and provides no
queue data and forces the skip to reduce further, leading to more rescheduling
and slihtly higher CPU use.  Now we only adjust the skip duration if there are
complete buffers or no data, not on fragments. I've set the initial skip lower
to help high bitrate streams.