handle the burst-size 0 case properly. shoult fix issue #28
some use that burst size or min queue size as 0 which was handled before the
queuing changes for rwlocks, The shrinking code for the no listener case went
too far and ending up leaving things inconsistent. now it does not go as far
and I don't need to maintain the count on the most recent buffer either.  I
also limit the amount I release each time when shrinking, that way a big
reduction from a large queue does not spend too long with a write lock.