fix possible bad pointer case
The original method was move to next buffer if at end of current buffer then write
which worked well when the queue blocks were refcounted but now they are not and
it is possible then the lag == queue size could refer to the block that is about to
go which would be a problem if the new block added is the same size.

So now, we write then check the position in the block and move to next, meaning that
we are always at the beginning of the new block by the end of the function. This of
course require a special case for lag == 0 as you cannot move forward in such cases.