fix for possible crash with range and on-demand relays
If you have an on-demand relay that has not been used since created and a listener connects
with a range request then the code path allowed them to get to the write route without a
per-format structure allocating.   This still does the short 206 drop but resets any 200
response.  This forces the listener code to rebuild the response headers after the stream
is going.

This only occurred on new on-demand relays, as soon as just 1 listener kicks off the relay
then the details would be present and what was missing is filled out.