add explicit flag for moving data on possible frame boundaries
This should only be an issue on files for listeners that need more flexible
searching to avoid things like ID3 or junk data when aligning for frames.
Normal queue processing requires the prevention of data moving.

Save extra processing and avoid possibly dropping parts of intro or fallback
files when sending to listeners.