fix for possible race in stats clearing with fallback files.
A possible case of listeners falling back to a file and then being
overridden by a reconnecting mount exposes a case where stats could be
released and then referenced by different file nodes.

We still drop the file node from the cache on override but now break the
stats handle leaving the stats cleanup to release the stats. To allow for
this we now update most stats updates via the stats thread. This also
reduces the number of updates when there are lots of listeners on a
fallback file as they are only done once or twice not for every listener