small mpeg parsing update.
some minor internal changes to reduce structure size by a small amount, add
api to enquire stream settings. expand on the frame marker checker to avoid
rare cases (typically with files) and have a more sensible trigger for the
reporting of the details in stats. We now report the actual detected stream
settings as well as the details reported via the headers, may help spot
misconfigurations.