only drop listener count if listener has finally left
it is possible that a listener can be in the process of being moved from
a source which fails and needs to be re-added back on for processing to
be consistent.  The source lock is dropped to allow the move attempt so
there is a small window of opportunity for the source to continue to the
complete shutdown and yet the listener is added back on.

Here the detach routine, removed the listener from the tree but leaves
the actual count reduction to the caller only if the listener is not
being added back on.