do not use SO_REUSEADDR on windows, breaks the reload handling
MS defined SO_REUSEADDR differently to BSD and linux and have allowed some stupid
security issue on it for port stealing. They messed it up, added another option
which doesn't help here and advise not using this option.  Luckily the default
behaviour is acceptable.  I've also avoided the abort case which should not trigger
but if it does, it reports an error and skips the rest.