drop listener if exactly queue size bytes behind
in a lagging listener case, a lag of queue size bytes means either they are on the first byte
of the last block (a just recoverable case) or on the last byte of the block that has just
gone. The latter case is possible but can fault the application in such cases.