could drop listeners at start up
in the case where lag is queue size bytes behind, it either means that they are
on the oldest block at position 0 (a valid case) or at the end of the possibly
removed block after the queue.  Because of the queue shrinkage code, it is quite
likely that the removed block case triggers more.