stats send updates. not many people using this but the changes here are internal
The key issued fixed here is a possible race where a source could get started between the
stats being generated and the client being added.  now we register the client, so that stats
for new details are added as normal, but the client itself is inactive (therefore no sending)
until the full list is made, and prepended. only then is the client made active. adn woken up

we also avoid the explicit wakeup, it's not required and a manual wakeup on exit is triggered.
This will be more useful later with relays switching masters.