move override trigger later, when some valid data has arrived
To handle cases wehere some clients start up to feed an override mount but fail to send anything
(usually because there feeds have failed).  By delaying the override of listeners, gaps do not
appear in their feeds. In normal cases the trigger still occurs.

instead of the source_init being used to trigger the override, the first buffer to place on the
queue is used.