remove block length guard og 15k as it could trip parent
We still limit check the memove to avoid very large copies as the parent should trap such
cases. we care more concerned about the amount we move in here than the original block length
as we may only move a small amount.

This has come up because libcurl has been known to send 15k+ blocks to the callback and this
was tripping up the mpeg routine into looping as the find sync point was not actually finding
the point and failing the frame detection routine, causing a recheck and loop.