avoid possible bad pointer on race case
It is possible that on a relay restart, after the stats are flushed, and the slave thread
kicks in, that the souce stat handle gets marked for purging as if it was a virtual mount
point.  As relays are never marked in such a way and are explicitly removed then we need
to exclude this marking case.