make sure we provide an IP on mount_add triggers, also provide useragent
Case to watch for here is slow auth. If the auth processing is getting delayed for whatever
reason then we cannot rely on the actual client details being around (although it will be in
most cases). If that goes then we cannot rely on a reference to it within auth, so copy what
we need.  Currently IP is straightforward, but useragent is within the parser tree, which is
wasteful to copy that, If other details are to be added later then that may need revisiting.