use an average for new block sizing on mp3/aac streams.
To help in reducing memory copies, the new block sizing when processing the
partial frames can now use an average from the previous block and the anchor
of the queue block size.  This allows for a certain amount of flexability in
sizing (which varies with code frame sizes) and keeps memory copy work down
in the common case.