update build scripts for kh11
This release has been waiting around for too long. many people have already
been using it from github and just some minor stuff at the end to tidy up.
Most of this commit is down to the updates in auto tools but the nsis file
also has new paths to the libs which seems to be more focused on using DLLs
for all the libs.