drop FD 0 1 2 before source [dis]connect exec.
Most FDs are not exported to a child process, but it can leak, and there is no real
good way of handling it generically except doing huge range of FDs.  most FDs now
use close-on-exec so should not leak through but this stops stdin/out/err from
going through as well. The program being called should not assume anything about
these so this limits any unwanted communication.