avoid possible crash case. hard to trigger
if a client rejoins the queue, after a short write occurs on the metadata then the blocks are
copied (including metadata), but this routine still referenced the previous metadata handle
which can be invalid now, an abort was getting triggered typically.

because the prev metadata was just used as a trigger, then avoid using that to reference a
resume from short write. Instead use the block associated pointer or the blank metadata.