prevent illegal instruction dump on icecast shutdown.
It only shows on later glibc and really means a multiple unlock which is undefined per spec.
the free tree routine which is clearing out any virtual source stats is not taking a lock on
each node before relaing, so add a wrapper for each case, no clients are present at the point
this is run.