some leftovers from the chunked/keep-alive changes
Disable the keep alive setting for typical stream connection, as that is unlikely to
be used again for any other connections.

Force ogg connections to be 1.0, thereby disabling chunked transfers.  May need to be
handled slightly differently later if we want keep alive but that is unlikely and makes
current clients work again.