rework a previous commit for memory corruption on fallback failure
The previous patch worked for what it was targeted for but it does not handle all cases, so
push the test we did lower into source listener release and only for the source specific parts.

This means that the source handle is NULLd before a move, which if fails, still goes through
the release, updates global stats and calls auth, but does not do source accounting twice (the
original cause of the said memory corruption).

It did highlight an extra verification test in the source shutdown that we could do in case the
counts do get messed up.