use pattern expansion on mount redirect.
The previous behaviour was not really that good, a simple hack for redirecting dead mounts to
other servers.  It made certain assumptions about the name. Now you can have

<mount>
  <mount-name>/stream*</mount-name>
  <redirect>http://otherserver:8000/${mount}</redirect>
</mount>

the question really is whether to put in a backward compatible hack for the previous behaviour
eg redirect http://other:8000/ would append the stream automatically.