expand the range of the block sizes for non-ogg streams
auto block sizing is based on incoming bitrate which works well but in cases where we
have many listeners, with multiple read 2-3k blocks, the default send size triggers
multiple sends per listener in a single pass.    Currently all listeners on a stream
will get processing on each of those passes.

This should allow smaller blocks for the lower bitrate streams so that the the default
send size does not trigger multiple sends and therefore delays the worker pass, but still
allows for the block to expand to allow for very large bitrate streams.