worker pass adjustments. spread processing client work for normal case.
keep timer updated for millisecond lookup, linux is quick these days for that but generally
it is OS platform dependent.

process only odd numbered of active scheduled clients during pass,  that way there is a
reasonable spread of clients processed on each pass. We could do a client count limiter but
that would really need a re-arrangement of the queue to allow the one missing out to get
scheduled next time.

obviosuly a worker that is closing down bypasses this limiter so that clients are either moved
or terminated.