Subtle API change for the log content routine, fix leak case
The headers on log content request was getting leaked.

API change is really for the allocation tally routines. Because log/* is unaware for it, the
overridden routines trip over misaligned if the routine allocates the memory.  Now we call
it initially with NULL and return the size and the log locked allowing the parent to allocate
memory as it sees fit, then call it again with the space allocated then filled out, eventually
the log is unlocked.

Return code gives 3 states, -1 for error, 1 for locked with size returned and 0 for completed.